Use the correct formula for determining deconv output shape
The earlier formula had assumed that `pad` meant padding on the
deconvolution input. However, `pad` means the padding on the
deconvolution output that would make a convolution from the output to
the input the reverse of this deconvolution. Thus, the formula for
deconvolution is simply what you get when you reverse `o_h` and `i_h`
in the convolution formula and then solve for `o_h` (and similarly for
width).